fix: more consistent drawer
- Consistent animations
- Break up components a bit
- Performance optimizations
- Better width adjusting for longer map titles
- Groundwork for future adjustments
