+++ src/components/layout/drawer/Drawer.jsx
@@ -1,6 +1,6 @@
-import React from 'react'
+import * as React from 'react'
 import Clear from '@material-ui/icons/Clear'
-import { Drawer, Typography, Grid, IconButton } from '@material-ui/core'
+import { Drawer, Typography, Grid, IconButton, Box } from '@material-ui/core'
 
 import { useStatic } from '@hooks/useStore'
 import Actions from './Actions'
@@ -13,6 +13,7 @@ export default function Sidebar({ drawer, toggleDrawer, toggleDialog }) {
     },
     ui,
   } = useStatic.getState()
+  const [width, setWidth] = React.useState(300)
 
   return (
     <Drawer
@@ -21,18 +22,23 @@ export default function Sidebar({ drawer, toggleDrawer, toggleDialog }) {
       open={drawer}
       onClose={toggleDrawer(false)}
     >
-      <Grid container alignItems="center" justifyContent="center">
+      <Grid
+        container
+        alignItems="center"
+        style={{ flexWrap: 'nowrap' }}
+        ref={(ref) => setWidth(Math.max(300, ref?.clientWidth || 0))}
+      >
         <Grid
           item
-          xs={2}
           className="grid-item"
           style={{
             backgroundImage: 'url(/favicon.ico)',
             width: 32,
             height: 32,
+            margin: 10,
           }}
         />
-        <Grid item xs={8}>
+        <Grid item style={{ flexGrow: 1 }}>
           <Typography
             variant="h5"
             color="secondary"
@@ -45,21 +51,23 @@ export default function Sidebar({ drawer, toggleDrawer, toggleDialog }) {
             {title}
           </Typography>
         </Grid>
-        <Grid item xs={2}>
+        <Grid item style={{ flexGrow: 0 }}>
           <IconButton onClick={toggleDrawer(false)}>
             <Clear style={{ color: 'white' }} />
           </IconButton>
         </Grid>
       </Grid>
-      {Object.entries(ui).map(([category, value]) => (
-        <DrawerSection
-          key={category}
-          category={category}
-          value={value}
-          toggleDialog={toggleDialog}
-        />
-      ))}
-      {separateDrawerActions && <Actions />}
+      <Box width={width}>
+        {Object.entries(ui).map(([category, value]) => (
+          <DrawerSection
+            key={category}
+            category={category}
+            value={value}
+            toggleDialog={toggleDialog}
+          />
+        ))}
+        {separateDrawerActions && <Actions />}
+      </Box>
     </Drawer>
   )
 }
